refactor: reorganize TableData module
- make use of new errors,
- split up into more organized files
